# gateway/src/http/openapi.yaml
openapi: 3.0.3
info:
  title: AgentZero Gateway API
  description: |
    RESTful API for the AgentZero AI agent runtime.

    ## Overview
    AgentZero provides a multi-agent orchestration platform with support for:
    - Agent configuration and management
    - Session-based conversations with context preservation
    - Real-time event streaming via SSE
    - MCP (Model Context Protocol) server integration
    - Webhook integrations for external platforms

    ## Authentication
    Currently, the API does not require authentication. This will be added in a future release.

    ## Base URLs
    - HTTP API: `http://localhost:18791`
    - WebSocket: `ws://localhost:18790`
  version: 0.1.0
  contact:
    name: AgentZero
  license:
    name: MIT
servers:
  - url: http://localhost:18791
    description: Local development server
...
paths:
...
  /api/connectors/{id}:
    get:
      tags: [Connectors]
      summary: Get connector
      description: Get a connector by ID
      operationId: getConnector
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Connector details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorConfig'
        '404':
          description: Connector not found
    put:
      tags: [Connectors]
      summary: Update connector
      description: Update an existing connector's configuration
      operationId: updateConnector
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConnectorRequest'
      responses:
        '200':
          description: Connector updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorConfig'
        '404':
          description: Connector not found
    delete:
      tags: [Connectors]
      summary: Delete connector
      description: Remove a registered connector
      operationId: deleteConnector
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Connector deleted
        '404':
          description: Connector not found
...
components:
  schemas:
...
    ConnectorConfig:
      type: object
      properties:
        id:
          type: string
          description: Unique connector identifier
          example: gmail-bridge
        name:
          type: string
          description: Human-readable name
          example: Gmail Bridge
        transport:
          $ref: '#/components/schemas/ConnectorTransport'
        metadata:
          $ref: '#/components/schemas/ConnectorMetadata'
        enabled:
          type: boolean
          description: Whether the connector is enabled
          default: true
        outbound_enabled:
          type: boolean
          description: Whether outbound dispatch is enabled
          default: true
        inbound_enabled:
          type: boolean
          description: Whether inbound message reception is enabled
          default: true
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
    ConnectorTransport:
      type: object
      description: Transport configuration for connector communication
      required:
        - type
      discriminator:
        propertyName: type
      oneOf:
        - $ref: '#/components/schemas/HttpTransport'
        - $ref: '#/components/schemas/CliTransport'
        - $ref: '#/components/schemas/GrpcTransport'
        - $ref: '#/components/schemas/WebSocketTransport'
        - $ref: '#/components/schemas/IpcTransport'
    HttpTransport:
      type: object
      properties:
        type:
          type: string
          enum: [http]
        callback_url:
          type: string
          format: uri
          description: URL to POST responses to
          example: http://localhost:9001/callback
        method:
          type: string
          enum: [POST, PUT]
          default: POST
        headers:
          type: object
          additionalProperties:
            type: string
          description: Custom headers to include
        timeout_ms:
          type: integer
          description: Request timeout in milliseconds
          nullable: true
    CliTransport:
      type: object
      properties:
        type:
          type: string
          enum: [cli]
        command:
          type: string
          description: Command to execute
          example: /usr/bin/notify-send
        args:
          type: array
          items:
            type: string
          description: Command arguments
        env:
          type: object
          additionalProperties:
            type: string
          description: Environment variables
    GrpcTransport:
      type: object
      properties:
        type:
          type: string
          enum: [grpc]
        endpoint:
          type: string
          description: gRPC endpoint
        service:
          type: string
          description: Service name
        method:
          type: string
          description: Method name
    WebSocketTransport:
      type: object
      properties:
        type:
          type: string
          enum: [websocket]
        url:
          type: string
          format: uri
          description: WebSocket URL
    IpcTransport:
      type: object
      properties:
        type:
          type: string
          enum: [ipc]
        socket_path:
          type: string
          description: Unix socket path
    ConnectorMetadata:
      type: object
      properties:
        capabilities:
          type: array
          items:
            $ref: '#/components/schemas/ConnectorCapability'
          description: Available actions/capabilities
        resources:
          type: array
          items:
            $ref: '#/components/schemas/ConnectorResource'
          description: Queryable data endpoints for agents
        response_schemas:
          type: array
          items:
            $ref: '#/components/schemas/ConnectorResponseSchema'
          description: Named outbound payload schemas
        context:
          type: string
          nullable: true
          description: Free-form text injected into agent prompts
      additionalProperties: true
      description: Additional connector-specific data
    ConnectorCapability:
      type: object
      properties:
        name:
          type: string
          description: Capability name
          example: send_email
        schema:
          type: object
          description: JSON schema for payload
        description:
          type: string
          nullable: true
          description: Human-readable description
    ConnectorResource:
      type: object
      required:
        - name
        - uri
        - method
      properties:
        name:
          type: string
          description: Resource name
          example: contacts
        uri:
          type: string
          description: URI or URI template
          example: https://api.example.com/contacts/{id}
        description:
          type: string
          nullable: true
          description: What this resource provides
        method:
          type: string
          description: HTTP method (GET or POST)
          default: GET
          example: GET
        headers:
          type: object
          additionalProperties:
            type: string
          description: Custom headers for the resource request
        response_schema:
          type: object
          nullable: true
          description: JSON Schema for the response
    ConnectorResponseSchema:
      type: object
      required:
        - name
        - schema
      properties:
        name:
          type: string
          description: Schema name
          example: send_message
        schema:
          type: object
          description: JSON Schema object
        description:
          type: string
          nullable: true
          description: What this schema defines
...
    UpdateConnectorRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        transport:
          $ref: '#/components/schemas/ConnectorTransport'
        metadata:
          $ref: '#/components/schemas/ConnectorMetadata'
        enabled:
          type: boolean
          nullable: true
        outbound_enabled:
          type: boolean
          nullable: true
        inbound_enabled:
          type: boolean
          nullable: true
...